Separate fake AWS things from unit tests to make them shrter and easier to read
