+++ services/worker-manager/test/fake-aws.js
@@ -1,0 +1,46 @@
+const assert = require('assert');
+
+module.exports = {
+  EC2: {
+    runInstances: ({defaultLaunchConfig, TagSpecifications, UserData}) => launchConfig => {
+      assert.deepStrictEqual(launchConfig,
+        {
+          ...defaultLaunchConfig.launchConfig,
+          MinCount: launchConfig.MinCount, // this is estimator's functionality, no need to test this here
+          MaxCount: launchConfig.MaxCount,
+          TagSpecifications,
+          UserData: JSON.stringify(UserData).toString('base64'),
+        }
+      );
+
+      let Instances = [];
+      for (let i = 0; i < launchConfig.MinCount; i++) {
+        Instances.push({
+          InstanceId: `i-${i}`,
+          AmiLaunchIndex: '',
+          ImageId: launchConfig.ImageId,
+          InstanceType: launchConfig.InstanceType || 'm2.micro',
+          Architecture: 'x86',
+          Placement: {
+            AvailabilityZone: 'someregion',
+          },
+          PrivateIpAddress: '1.1.1.1',
+          OwnerId: '123123123',
+          State: {
+            Name: 'running',
+          },
+          StateReason: {
+            Message: 'yOu LaunCHed iT!!!1',
+          },
+        });
+      }
+      return {
+        promise: () => ({
+          Instances,
+          Groups: [],
+          OwnerId: '123123123',
+        }),
+      };
+    },
+  },
+};
